pyproject.toml
[project]
name = "play-pgvector"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "yamatatsu", email = "the.kinnikuman@gmail.com" }
]
dependencies = [
    "altair>=5.5.0",
    "anthropic>=0.42.0",
    "awscli>=1.36.30",
    "boto3-stubs>=1.35.90",
    "boto3>=1.35.89",
    "botocore>=1.35.89",
    "langchain-aws<=0.2.1",
    "langchain-community>=0.3.13",
    "langchain-postgres>=0.0.12",
    "langchain>=0.3.13",
    "pgvector<0.3.0",
    "psycopg[binary,pool]>=3.2.3",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
    "types-psycopg2>=2.9.21.20241019",
    "urllib3>=2.3.0",
    "watchdog>=6.0.0",
    "langgraph>=0.2.60",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye.scripts]
play_pgvector = "streamlit run src/app.py"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/play_pgvector"]
